[Aiu팀]Flask서버_코드/requirements.txt
flask
python-dotenv
requests
mysql-connector-python
tensorflow
numpy
pandas
scikit-learn
